- brats 2019 (seg)

+++ datasets.py
@@ -536,7 +536,7 @@ class WeightedSubsetRandomSampler(Sampler):
     def __init__(self, num_samples, sample_div=2):
         self.num_samples = num_samples // sample_div
         self.weights = torch.tensor(
-            [np.iinfo(np.double).max] * num_samples, dtype=torch.double
+            [np.iinfo(np.long).max] * num_samples, dtype=torch.double
         )
 
     def __iter__(self):
